Passing `extension_options` values to the `register` function of zope.sqlalchemy
The session status remains `active` if the extension_options values are not passed to the `register` function of `zope.sqlalchemy`.

This leads to the transaction rollback on cursor.execute operations.

The session status should be marked as `changed` after the cursor.execute operation for the transaction to commit.

This is done by passing
`extension_options={'initial_state': 'invalidated'}`

https://github.com/zopefoundation/Products.SQLAlchemyDA/blob/2.0/src/Products/SQLAlchemyDA/da.py#L261

The z3c.sqlalchemy should send the `extension_options` to the `register` function of `zope.sqlalchemy`

This fix should go to `z3c.sqlalchemy > 1.4.0`